+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/http/service/PresentProofApiServiceImpl.scala
@@ -8,10 +8,10 @@ import akka.http.scaladsl.server.Route
 import zio._
 import scala.concurrent.Future
 import io.iohk.atala.agent.server.http.model.HttpServiceError
+import io.iohk.atala.mercury._
 import io.iohk.atala.mercury.model.DidId
 import io.iohk.atala.mercury.protocol.presentproof._
 import java.util.UUID
-import io.iohk.atala.mercury.DidComm
 import io.iohk.atala.agent.server.http.model.OASDomainModelHelper
 import io.iohk.atala.agent.server.http.model.OASErrorModelHelper
 import io.iohk.atala.agent.server.http.model.InvalidState
@@ -28,7 +28,6 @@ import io.iohk.atala.mercury.model.Base64
 class PresentProofApiServiceImpl(
     presentationService: PresentationService,
     connectionService: ConnectionService,
-    didCommService: DidComm
 )(using runtime: Runtime[Any])
     extends PresentProofApiService
     with AkkaZioSupport
@@ -162,13 +161,12 @@ class PresentProofApiServiceImpl(
 }
 
 object PresentProofApiServiceImpl {
-  val layer: URLayer[PresentationService & ConnectionService & DidComm, PresentProofApiService] = ZLayer.fromZIO {
+  val layer: URLayer[PresentationService & ConnectionService, PresentProofApiService] = ZLayer.fromZIO {
 
     for {
       rt <- ZIO.runtime[Any]
       presentationService <- ZIO.service[PresentationService]
       connectionService <- ZIO.service[ConnectionService]
-      didCommService <- ZIO.service[DidComm]
-    } yield PresentProofApiServiceImpl(presentationService, connectionService, didCommService)(using rt)
+    } yield PresentProofApiServiceImpl(presentationService, connectionService)(using rt)
   }
 }