+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/jobs/BackgroundJobs.scala
@@ -11,10 +11,7 @@ import zio.Duration
 import java.time.Instant
 import java.time.Clock
 import java.time.ZoneId
-import io.iohk.atala.mercury.DidComm
-import io.iohk.atala.mercury.MediaTypes
-import io.iohk.atala.mercury.MessagingService
-import io.iohk.atala.mercury.HttpClient
+import io.iohk.atala.mercury._
 import io.iohk.atala.mercury.model._
 import io.iohk.atala.mercury.model.error._
 import io.iohk.atala.mercury.protocol.issuecredential._
@@ -30,12 +27,13 @@ import io.iohk.atala.pollux.core.model.PresentationRecord
 import io.iohk.atala.pollux.core.service.PresentationService
 import io.iohk.atala.pollux.core.model.error.PresentationError
 import io.iohk.atala.agent.server.http.model.{InvalidState, NotImplemented}
-import io.iohk.atala.agent.walletapi.service.ManagedDIDService
-import io.iohk.atala.mercury.AgentServiceAny
 import org.didcommx.didcomm.DIDComm
-import io.iohk.atala.agent.walletapi.model.error.DIDSecretStorageError
-import io.iohk.atala.agent.walletapi.model.ManagedDIDTemplate
+import io.iohk.atala.agent.walletapi.model._
+import io.iohk.atala.agent.walletapi.model.error._
+import io.iohk.atala.agent.walletapi.model.error.DIDSecretStorageError.KeyNotFoundError
+import io.iohk.atala.agent.walletapi.service.ManagedDIDService
 import io.iohk.atala.agent.walletapi.sql.JdbcDIDSecretStorage
+import io.iohk.atala.agent.walletapi.storage.DIDSecretStorage
 import io.iohk.atala.pollux.vc.jwt.ES256KSigner
 import io.iohk.atala.castor.core.model.did._
 import java.security.KeyFactory
@@ -51,11 +49,8 @@ import java.security.spec.ECPoint
 import org.bouncycastle.jce.provider.BouncyCastleProvider
 import io.circe.Json
 import io.circe.syntax._
-import io.iohk.atala.agent.walletapi.storage.DIDSecretStorage
-import io.iohk.atala.agent.walletapi.model.error.CreateManagedDIDError
 import io.iohk.atala.pollux.vc.jwt.JWT
 import io.iohk.atala.pollux.vc.jwt.{DidResolver => JwtDidResolver}
-import io.iohk.atala.agent.walletapi.model._
 
 object BackgroundJobs {
 
@@ -81,7 +76,7 @@ object BackgroundJobs {
   private[this] def performExchange(
       record: IssueCredentialRecord
   ): URIO[
-    DIDResolver & JwtDidResolver & HttpClient & CredentialService & ManagedDIDService & DIDSecretStorage,
+    DidOps & DIDResolver & JwtDidResolver & HttpClient & CredentialService & ManagedDIDService & DIDSecretStorage,
     Unit
   ] = {
     import IssueCredentialRecord._
@@ -95,7 +90,9 @@ object BackgroundJobs {
           for {
             _ <- ZIO.log(s"IssueCredentialRecord: OfferPending (START)")
             didCommAgent <- buildDIDCommAgent(offer.from)
-            _ <- MessagingService.send(offer.makeMessage).provideSomeLayer(didCommAgent)
+            _ <- MessagingService
+              .send(offer.makeMessage)
+              .provideSomeLayer(didCommAgent)
             credentialService <- ZIO.service[CredentialService]
             _ <- credentialService.markOfferSent(id)
           } yield ()
@@ -121,7 +118,9 @@ object BackgroundJobs {
             ) =>
           for {
             didCommAgent <- buildDIDCommAgent(request.from)
-            _ <- MessagingService.send(request.makeMessage).provideSomeLayer(didCommAgent)
+            _ <- MessagingService
+              .send(request.makeMessage)
+              .provideSomeLayer(didCommAgent)
             credentialService <- ZIO.service[CredentialService]
             _ <- credentialService.markRequestSent(id)
           } yield ()
@@ -186,7 +185,7 @@ object BackgroundJobs {
               fromDID = issue.from,
               toDID = issue.to,
               thid = issue.thid,
-              credentials = Map("prims/jwt" -> signedJwtCredential.value)
+              credentials = Map("prims/jwt" -> signedJwtCredential.value.getBytes)
             )
             _ <- credentialService.markCredentialGenerated(id, issueCredential)
 
@@ -213,7 +212,9 @@ object BackgroundJobs {
             ) =>
           for {
             didCommAgent <- buildDIDCommAgent(issue.from)
-            _ <- MessagingService.send(issue.makeMessage).provideSomeLayer(didCommAgent)
+            _ <- MessagingService
+              .send(issue.makeMessage)
+              .provideSomeLayer(didCommAgent)
             credentialService <- ZIO.service[CredentialService]
             _ <- credentialService.markCredentialSent(id)
           } yield ()
@@ -315,7 +316,7 @@ object BackgroundJobs {
   private[this] def performPresentation(
       record: PresentationRecord
   ): URIO[
-    DIDResolver & JwtDidResolver & HttpClient & PresentationService & ManagedDIDService & DIDSecretStorage,
+    DidOps & DIDResolver & JwtDidResolver & HttpClient & PresentationService & ManagedDIDService & DIDSecretStorage,
     Unit
   ] = {
     import io.iohk.atala.pollux.core.model.PresentationRecord.ProtocolState._
@@ -337,6 +338,7 @@ object BackgroundJobs {
             case Some(record) =>
               for {
                 _ <- ZIO.log(s"PresentationRecord: RequestPending (Send Massage)")
+                didOps <- ZIO.service[DidOps]
                 didCommAgent <- buildDIDCommAgent(record.from)
                 _ <- MessagingService.send(record.makeMessage).provideSomeLayer(didCommAgent)
                 service <- ZIO.service[PresentationService]
@@ -388,7 +390,10 @@ object BackgroundJobs {
                     ),
                     attachments = Seq(
                       AttachmentDescriptor
-                        .buildAttachment(payload = signedJwtPresentation.value, mediaType = Some("prism/jwt"))
+                        .buildBase64Attachment(
+                          payload = signedJwtPresentation.value.getBytes(),
+                          mediaType = Some("prism/jwt")
+                        )
                     ),
                     thid = requestPresentation.thid.orElse(Some(requestPresentation.id)),
                     from = requestPresentation.to,
@@ -407,7 +412,9 @@ object BackgroundJobs {
               for {
                 _ <- ZIO.log(s"PresentationRecord: PresentationPending (Send Message)")
                 didCommAgent <- buildDIDCommAgent(p.from)
-                _ <- MessagingService.send(p.makeMessage).provideSomeLayer(didCommAgent)
+                _ <- MessagingService
+                  .send(p.makeMessage)
+                  .provideSomeLayer(didCommAgent)
                 service <- ZIO.service[PresentationService]
                 _ <- service.markPresentationSent(id)
               } yield ()
@@ -424,7 +431,7 @@ object BackgroundJobs {
                 didResolverService <- ZIO.service[JwtDidResolver]
                 credentialsValidationResult <- p.attachments.head.data match {
                   case Base64(data) =>
-                    val base64Decoded = new String(java.util.Base64.getDecoder().decode(data)).drop(1).dropRight(1)
+                    val base64Decoded = new String(java.util.Base64.getDecoder().decode(data))
 
                     println(s"Base64decode:\n\n ${base64Decoded} \n\n")
                     JwtPresentation.verify(
@@ -474,17 +481,23 @@ object BackgroundJobs {
       }
   }
 
-  private[this] def buildDIDCommAgent(myDid: DidId) = {
+  // private[this] def buildDIDCommAgent(myDid: DidId): ZLayer[ManagedDIDService, KeyNotFoundError, DidAgent] = { // FIXME
+  //   val aux = for {
+  //     managedDidService <- ZIO.service[ManagedDIDService]
+  //     peerDID <- managedDidService.getPeerDID(myDid)
+  //     agent = AgentPeerService.makeLayer(peerDID)
+  //   } yield agent
+  //   ZLayer.fromZIO(aux).flatten
+  // }
+
+  private[this] def buildDIDCommAgent(
+      myDid: DidId
+  ): ZIO[ManagedDIDService, KeyNotFoundError, ZLayer[Any, Nothing, DidAgent]] = {
     for {
       managedDidService <- ZIO.service[ManagedDIDService]
       peerDID <- managedDidService.getPeerDID(myDid)
-      didCommAgent = ZLayer.succeed(
-        AgentServiceAny(
-          new DIDComm(UniversalDidResolver, peerDID.getSecretResolverInMemory),
-          peerDID.did
-        )
-      )
-    } yield didCommAgent
+      agent = AgentPeerService.makeLayer(peerDID)
+    } yield agent
   }
 
   val publishCredentialsToDlt = {